working extrapol layers (performance test pending)
Appended extrapolation layers to the registration layers. Modified the
training and model definition to first pre-train the extrapolation
layers to approximate identity (otherwise it produces nonsense).

Also added config option (on a per stage level) to enable/disable
trainability of registration layers and of extrapolation layers (each
separate config option). The algorithm seems to be working when the
first couple of stages are run with only registration layers with ADAM
optimizer, and then extrapolation layers are switched on adn SGD is
used.